Move getBackgroundImage function into backgrounds store
